requirements.txt
Flask
gunicorn
vertexai==1.71.1
google-cloud-aiplatform==1.71.1
google-cloud-bigquery
pydantic
Flask-Cors
flask-dance
bcrypt